client/requirements.txt
requests
pyroute2
docker-py
